Lesson_7 completed

+++ Lesson_7/view/tasks.php
@@ -7,36 +7,45 @@
 </style>
 
 <body>
-    <div class="container">
-        <div class="topContainer">
-            <div class="menu">
-                <a href="/">Main</a>
-                <a href="/?controller=addTask">Add Task</a>
-                <p>Hello <?= $userName ?>.</p>
-            </div>
-            <div class="header">Task controller</div>
+    <div class="topContainer">
+        <div class="menu">
+            <a href="/">Main</a>
+            <a href="/?controller=addTask">Add Task</a>
+            <p>Hello <?= $userName ?>.</p>
         </div>
+        <div class="header">Task controller</div>
+    </div>
 
-        <div class="tasks">
-            <p>Task List</p>
-            <div class="taskHeader">
-                <p>Description</p>
-                <p>Priority</p>
-                <p>Mark as done</p>
-            </div>
-            <?php if ($tasks !== null) : ?>
-                <?php foreach ($tasks as $key => $task) : ?>
-                    <div class="taskList">
-                        <p><?= $task->getDescription() ?></p>
-                        <p><?= $task->getPriority() ?></p>
+    <a href='/?action=showCompleted'>
+        <?php if ($status == null || $status == 0) : ?>
+            <button class="showBtn">Show completed tasks</button>
+        <?php else : ?>
+            <button class="showBtn">Show incompleted tasks</button>
+        <?php endif; ?>
+    </a>
+
+    <div class="tasks">
+        <p>Task List</p>
+        <div class="taskHeader">
+            <p>Description</p>
+            <p>Priority</p>
+            <p>Mark as done</p>
+        </div>
+        <?php if ($tasks !== null) : ?>
+            <?php foreach ($tasks as $key => $task) : ?>
+                <div class="taskList">
+                    <p><?= $task->getDescription() ?></p>
+                    <p><?= $task->getPriority() ?></p>
+                    <?php if ($status == null || $status == 0) : ?>
                         <a href="/?action=delete&key=<?= $task->getId() ?>">[Task done]</a>
-                    </div>
-                <?php endforeach; ?>
-            <?php else : ?>
-                <p>No active tasks</p>
-            <?php endif; ?>
-        </div>
-
+                    <?php else : ?>
+                        <a href="/?action=incomplete&key=<?= $task->getId() ?>">[Task incomplete]</a>
+                    <?php endif; ?>
+                </div>
+            <?php endforeach; ?>
+        <?php else : ?>
+            <p>No active tasks</p>
+        <?php endif; ?>
     </div>
 
 </body>
